Fix bug in initialization of allowed values (#43)
This fixes #42.

The bug was in copying reference of array containing values from config into array containing all allowed values. The latter was then modified (by adding values and sorting), which caused items in the values array to change.

This PR replaces copy by reference to creation of new array and coping values and modifies test to go through the previously failing scenario.