Add detach support which is compatible with hcl2 used for terraform 0.12 (#45)
This fixes #44

Context:
--detach flag is used in astro to make terraform migrations easier. When terraform resources need to be moved or deleted terraform state subcommands are used to manipulate the state file.
If the remote backend is configured, state file will be uploaded after state manipulation is successful. Which is not desired while migration script is being tested. With detach option astro inits the module (this downloads the state file from remote), modifies the configuration to remove remote backend (which makes terraform use "local" backend) and runs terraform init -force-copy again to re-initialize with new backend. See more in https://www.terraform.io/docs/backends/config.html#unconfiguring-a-backend.

There were two issues with --detach flag that were uncovered by terraform 0.12 upgrade:

astro failed if there was a file in the config that contained terraform block, but didn't contain backend block. This bug exists with the older versions of terraform. This is fixed by changing astDel to astDelIfExists.
To remove remote backend from config, astro walks and modifies the HCL AST tree.
Terraform 0.12 switched to the new language (HCL2), so old parsing mechanism stopped working.
I wasn't able to find a way to walk and modify tree with https://godoc.org/github.com/hashicorp/hcl2 (similar issue reported hashicorp/hcl2#23), so I'm using regexp to do this for the simple configurations. That unblocks us for the terraform 0.12 upgrade, but may be not a good solution for the long term.