Add release target to Makefile (#62)
The approach that was used before to build release was involving using goreleaser docker container (see Dockerfile).
The problem with that was that Docker container uses Linux Alpine, which has the different version of libc. Some dependency of astro uses libc, so with the previous approach the binary won't start because libc.musl-x86_64.so.1 is not available on Debian by default.

The new approach is to start golang docker container and make a release there.